services/graph-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
neo4j==5.16.0
networkx==3.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
httpx==0.26.0
